feat(server:auth): allow cookie to contain token
This enables users to access the API from the browser without having to
set headers.

If the setTokenCookie is added for all authentication strategies, it
also makes API tests simpler, since they can just use a request agent
with a cookie store (See #494).

Note the wrapping if that only does this when there's no Authorization
header.