feat(gen) add Dockerfile and docker-compose.yml support
Add support to run generated Angular-Fullstack applications as Docker containers by including
Dockerfile, docker-compose.yml and .dockerignore. When the application is built with grunt
build, these files are copied to the dist directory. The Docker container can then be built
and started there with the "docker-compose up" command. When Mongoose is used, Docker Compose
will also launch a MongoDB container, eliminating the need to install MongoDB on the local
machine. Note that Docker and Docker Compose must be installed for Docker support to work.
They are available on Linux, OS X and Windows.