Error handling for email not provided
Many times, Facebook users will not allow the app to get the email. If they do this, the app will literally crash and they will see just a white screen with nowhere else to go. This change will let the app check wether the email field was provided, and if it wasn't, it'll just make an empty string of it. Would anyone else suggest a different datatype?