fix(user): fix email and password validation
If no email and password were passed to the authentication endpoints,
the authentication would still succeed even though there was no email
and password. This fixes that to make sure that we account for blank and
null email and passwords. Also it makes sure that email
and password are not required when using other oauth providers.