fix: build images before rev-replace
When running `gulp build` the rev manifest will be remade, but the task
"build:client" which uses this file doesn't depend on "build:images"
which populates this file. The result is that when running `build`,
rev-replace will be unable to replace images as it doesn't wait for
images to be filled in.
Solution is to simply have "build:client" depend upon "build:images"