fix(seed): remove return from Thing.destroy
The return on Thing.destroy was preventing the creation of the users when filters.sequelizeModels and filters.auth were both true. A better way to handle them would be to import sequelize then wrap the creation of things and users in Sequelize.Promise.all. That would prevent the warnings for unhandled promises. Just wanted to get the bug fixed for now. Can work on trying to make it prettier later.